requirements.txt
flask
matplotlib
numpy
pillow
python-dotenv
pytest
torch
